+++ protobuf/lib/src/protobuf/builder_info.dart
@@ -160,6 +160,18 @@ class BuilderInfo {
         valueOf, enumValues);
   }
 
+  void aOM<T extends GeneratedMessage>(
+      int tagNumber, String name, T Function() create) {
+    add<T>(tagNumber, name, PbFieldType.OM,
+        GeneratedMessage._defaultMakerFor<T>(create), create, null, null);
+  }
+
+  void aQM<T extends GeneratedMessage>(
+      int tagNumber, String name, T Function() create) {
+    add<T>(tagNumber, name, PbFieldType.QM,
+        GeneratedMessage._defaultMakerFor<T>(create), create, null, null);
+  }
+
   // oneof declarations.
   void oo(int oneofIndex, List<int> tags) {
     tags.forEach((int tag) => oneofs[tag] = oneofIndex);