+++ protoc_plugin/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 18.0.3
+
+* Add specialized getters for `String`, `int`, and `bool` with usual default values.
+* Shrink dart2js generated code for `getDefault()`.
+
 ## 18.0.2
 
 * Fix mangling of extension names, message type names, and enum names that are Dart keywords.